+++ src/main/kotlin/com/lizhuolun/feignhelper/cache/CacheChangeNotifier.kt
@@ -1,0 +1,28 @@
+package com.lizhuolun.feignhelper.cache
+
+import com.intellij.util.messages.Topic
+
+/**
+ * 缓存变更监听器，用于解耦缓存层与 UI 层。
+ *
+ * @author lizhuolun
+ * @date 2026/6/12
+ */
+interface CacheChangeListener {
+
+    /**
+     * 双边映射缓存发生变更时回调。
+     */
+    fun onCacheChanged()
+
+    companion object {
+
+        /**
+         * 消息总线 topic。
+         */
+        val TOPIC: Topic<CacheChangeListener> = Topic.create(
+            "FeignHelper Cache Changed",
+            CacheChangeListener::class.java,
+        )
+    }
+}